Added deep feature extraction to STORN

+++ greenarm/models/sampling/sampling.py
@@ -4,6 +4,7 @@ import keras.backend as K
 def sample_gauss(mu, sig, batch_size, dim_size):
     epsilon = K.random_normal(shape=(batch_size,
                                      dim_size),
-                              mean=0., std=1. )
+                              mean=0., std=1.,
+                              dtype="float32")
     sample = mu + sig * epsilon
     return sample
